Update WindowsScalability.ini
A value of 0 for r.SkeletalMeshLODBias helps fix the ugly NPC pop-in. A setting of 4 is worse than vanilla's 2 but it might give 1-4fps more on lower -end systems.

For some reason, the Background Model Detail's Medium setting is mapped to [StaticMeshQuality@3].

The r.MassiveEnvironment.ViewLODScale setting is what ultimately causes the terrain pop-in when it's set to values < 1.